+++ packages/aesthetic-adapter-aphrodite/src/UnifiedAdapter.js
@@ -13,7 +13,7 @@ import type { StyleDeclarationMap, TransformedStylesMap, CSSStyle } from '../../
 export default class UnifiedAphroditeAdapter extends AphroditeAdapter {
   syntax: UnifiedSyntax;
 
-  constructor(aphrodite: Object, options: Object = {}) {
+  constructor(aphrodite: Object, options?: Object = {}) {
     super(aphrodite, options);
 
     this.syntax = new UnifiedSyntax();