+++ packages/aesthetic-adapter-jss/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 🛠 Internal
 * Enabled Yarn workspaces.
+* Updated Flowtype definitions.
 
 # 1.6.0 - 9/27/17
 #### 🚀 New